Problem: tracking and discussing proposals
Some changes are too important or complicated to be just
merged in. However, it is hard to track and discuss these
changes because the discussions and decisions are captured
in various project issues. Finalized decisions are often
not recorded well for consumption furher down the road.

Solution: designate a repository of RFCs (Requests For Comments)
where the decisions are formalized in a form of documents.